Remove duplication in climate group forms
