+++ src/main/java/uk/co/fivium/els/categories/airconditioners/model/HeatingDuctlessAirConditionersForm.java
@@ -1,15 +1,12 @@
 package uk.co.fivium.els.categories.airconditioners.model;
 
-import org.hibernate.validator.group.GroupSequenceProvider;
-import uk.co.fivium.els.categories.common.StandardTemplateForm50Char;
-import uk.co.fivium.els.model.meta.FieldPrompt;
-
 import javax.validation.constraints.Digits;
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import org.hibernate.validator.group.GroupSequenceProvider;
+import uk.co.fivium.els.model.meta.FieldPrompt;
 
 @GroupSequenceProvider(HeatingDuctlessAirConditionersFormSequenceProvider.class)
-public class HeatingDuctlessAirConditionersForm extends StandardTemplateForm50Char {
+public class HeatingDuctlessAirConditionersForm extends MultipleClimateGroupForm {
 
   @FieldPrompt("Energy efficiency class for Average heating season")
   @NotBlank(message = "Select an energy efficiency indicator for average climate conditions")
@@ -26,46 +23,6 @@ public class HeatingDuctlessAirConditionersForm extends StandardTemplateForm50Ch
   @FieldPrompt("Annual energy consumption in kWh per year in average climate conditions")
   @Digits(integer = 4, fraction = 0,message = "Enter the annual energy consumption for average climate conditions, up to 4 digits")
   private String averageAnnualEnergyConsumption;
-
-  @FieldPrompt("Is data available for warmer climate conditons?")
-  @NotNull(message = "Select whether data is available for warmer climate conditions")
-  private Boolean warmerConditions;
-
-  @FieldPrompt("Energy efficiency class for Warmer heating season")
-  @NotBlank(groups = WarmerClimateGroup.class, message = "Select an energy efficiency indicator for warmer climate conditions")
-  private String warmerHeatingEfficiencyRating;
-
-  @FieldPrompt("Design load for heating in warmer climate conditions in kW")
-  @Digits(groups = WarmerClimateGroup.class, integer = 2, fraction = 1, message = "Enter the design load for warmer climate conditions, up to 2 digits with an optional decimal place")
-  private String warmerHeatingDesignLoad;
-
-  @FieldPrompt("Seasonal coefficient of performance in warmer climate conditions (SCOP value)")
-  @Digits(groups = WarmerClimateGroup.class, integer = 1, fraction = 1, message = "Enter the SCOP value for warmer climate conditions, 1 digit with an optional decimal place")
-  private String warmerScop;
-
-  @FieldPrompt("Annual energy consumption in kWh per year in warmer climate conditions")
-  @Digits(groups = WarmerClimateGroup.class, integer = 4, fraction = 0,message = "Enter the annual energy consumption for warmer climate conditions, up to 4 digits")
-  private String warmerAnnualEnergyConsumption;
-
-  @FieldPrompt("Is data available for colder climate conditons?")
-  @NotNull(message = "Select whether data is available for colder climate conditions")
-  private Boolean colderConditions;
-
-  @FieldPrompt("Energy efficiency class for Colder heating season")
-  @NotBlank(groups = ColderClimateGroup.class, message = "Select an energy efficiency indicator for colder climate conditions")
-  private String colderHeatingEfficiencyRating;
-
-  @FieldPrompt("Design load for heating in colder climate conditions in kW")
-  @Digits(groups = ColderClimateGroup.class, integer = 2, fraction = 1, message = "Enter the design load for colder climate conditions, up to 2 digits with an optional decimal place")
-  private String colderHeatingDesignLoad;
-
-  @FieldPrompt("Seasonal coefficient of performance in colder climate conditions (SCOP value)")
-  @Digits(groups = ColderClimateGroup.class, integer = 1, fraction = 1, message = "Enter the SCOP value for colder climate conditions, 1 digit with an optional decimal place")
-  private String colderScop;
-
-  @FieldPrompt("Annual energy consumption in kWh per year in colder climate conditions")
-  @Digits(groups = ColderClimateGroup.class, integer = 4, fraction = 0,message = "Enter the annual energy consumption for colder climate conditions, up to 4 digits")
-  private String colderAnnualEnergyConsumption;
 
   @FieldPrompt("Sound power levels for indoor units expressed in dB(A) re1 pW")
   @Digits(integer = 2, fraction = 0, message = "Enter the outdoors sound power level, up to 2 digits long")
@@ -105,86 +62,6 @@ public class HeatingDuctlessAirConditionersForm extends StandardTemplateForm50Ch
 
   public void setAverageAnnualEnergyConsumption(String averageAnnualEnergyConsumption) {
     this.averageAnnualEnergyConsumption = averageAnnualEnergyConsumption;
-  }
-
-  public Boolean getWarmerConditions() {
-    return warmerConditions;
-  }
-
-  public void setWarmerConditions(Boolean warmerConditions) {
-    this.warmerConditions = warmerConditions;
-  }
-
-  public String getWarmerHeatingEfficiencyRating() {
-    return warmerHeatingEfficiencyRating;
-  }
-
-  public void setWarmerHeatingEfficiencyRating(String warmerHeatingEfficiencyRating) {
-    this.warmerHeatingEfficiencyRating = warmerHeatingEfficiencyRating;
-  }
-
-  public String getWarmerHeatingDesignLoad() {
-    return warmerHeatingDesignLoad;
-  }
-
-  public void setWarmerHeatingDesignLoad(String warmerHeatingDesignLoad) {
-    this.warmerHeatingDesignLoad = warmerHeatingDesignLoad;
-  }
-
-  public String getWarmerScop() {
-    return warmerScop;
-  }
-
-  public void setWarmerScop(String warmerScop) {
-    this.warmerScop = warmerScop;
-  }
-
-  public String getWarmerAnnualEnergyConsumption() {
-    return warmerAnnualEnergyConsumption;
-  }
-
-  public void setWarmerAnnualEnergyConsumption(String warmerAnnualEnergyConsumption) {
-    this.warmerAnnualEnergyConsumption = warmerAnnualEnergyConsumption;
-  }
-
-  public Boolean getColderConditions() {
-    return colderConditions;
-  }
-
-  public void setColderConditions(Boolean colderConditions) {
-    this.colderConditions = colderConditions;
-  }
-
-  public String getColderHeatingEfficiencyRating() {
-    return colderHeatingEfficiencyRating;
-  }
-
-  public void setColderHeatingEfficiencyRating(String colderHeatingEfficiencyRating) {
-    this.colderHeatingEfficiencyRating = colderHeatingEfficiencyRating;
-  }
-
-  public String getColderHeatingDesignLoad() {
-    return colderHeatingDesignLoad;
-  }
-
-  public void setColderHeatingDesignLoad(String colderHeatingDesignLoad) {
-    this.colderHeatingDesignLoad = colderHeatingDesignLoad;
-  }
-
-  public String getColderScop() {
-    return colderScop;
-  }
-
-  public void setColderScop(String colderScop) {
-    this.colderScop = colderScop;
-  }
-
-  public String getColderAnnualEnergyConsumption() {
-    return colderAnnualEnergyConsumption;
-  }
-
-  public void setColderAnnualEnergyConsumption(String colderAnnualEnergyConsumption) {
-    this.colderAnnualEnergyConsumption = colderAnnualEnergyConsumption;
   }
 
   public String getSoundPowerLevelIndoors() {
